bcachefs: Rework of cut_front & cut_back
Mainline: 085ab69357e091613625f1505d667b6a5a3ec881
From: v6.7-rc1
Severity: Moderate

task: 205710

This changes bch2_cut_front and bch2_cut_back so that they're able to
shorten the size of the value, and it also changes the extent update
path to update the accounting in the btree node when this happens.

When the size of the value is shortened, they zero out the space that's
no longer used, so it's interpreted as noops (as implemented in the last
patch).

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>
Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>
(backported from commit 085ab69357e091613625f1505d667b6a5a3ec881)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: I9d4a24f678d228421248e8a6d3919c644258d24b