bcachefs: Fix assertion popping in transaction commit path
Mainline: aa8889c07abecd7db7b2c0beb61db921fbafe04f
From: v6.7-rc1
Severity: Moderate

task: 205710

We can't be holding read locks on btree nodes when we go to take write
locks: this would deadlock if another thread is holding an intent lock
on the node we have a read lock on, and it tries to commit and upgrade
to a write lock.

But instead of triggering an assertion, if this happens we can just
upgrade the read lock to an intent lock.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>
Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>
(backported from commit aa8889c07abecd7db7b2c0beb61db921fbafe04f)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: I67a1a4171dbadcf4a6f602c01554a85393228d5c