bcachefs: KEY_TYPE_alloc_v2
Mainline: 7f4e1d5d0faff0d72e9f6708bf98488d76533846
From: v6.7-rc1
Severity: Moderate

task: 205710

This introduces a new version of KEY_TYPE_alloc, which uses the new
varint encoding introduced for inodes. This means we'll eventually be
able to support much larger bucket sizes (for SMR devices), and the
read/write time fields are expanded to 64 bits - which will be used in
the next patch to get rid of the periodic rescaling of those fields.

Also, for buckets that are members of erasure coded stripes, this adds
persistent fields for the index of the stripe they're members of and the
stripe redundancy. This is part of work to get rid of having to scan and
read into memory the alloc and stripes btrees at mount time.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>
Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>
(backported from commit 7f4e1d5d0faff0d72e9f6708bf98488d76533846)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: I5e929d8743417d0564eae1c49e623f509365442a