bcachefs: Kill bch2_btree_iter_set_pos_same_leaf()
Mainline: 792e2c4c850c75857e822fa40660a3b4733d5b8c
From: v6.7-rc1
Severity: Moderate

task: 205710

The only reason we were keeping this around was for
BTREE_INSERT_NOUNLOCK semantics - if bch2_btree_iter_set_pos() advances
to the next leaf node, it'll drop the lock on the node that we just
inserted to.

But we don't rely on BTREE_INSERT_NOUNLOCK semantics for the extents
btree, just the inodes btree, and if we do need it for the extents btree
in the future we can do it more cleanly by cloning the iterator - this
lets us delete some special cases in the btree iterator code, which is
complicated enough as it is.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>
Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>
(backported from commit 792e2c4c850c75857e822fa40660a3b4733d5b8c)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: I0cf72fbe0874c8267eea812f81946d0342da244f