OPENEULER: net: hns3: add support for queue bonding mode of flow director
Mainline: KYLIN-only
Severity: Moderate

task: 218995

For device version V3, it supports queue bonding, which can
identify the tuple information of TCP stream, and create flow
director rules automatically, in order to keep the tx and rx
packets are in the same queue pair. The driver set FD_ADD
field of TX BD for TCP SYN packet, and set FD_DEL filed for
TCP FIN or RST packet. The hardware create or remove a fd rule
according to the TX BD, and it also support to age-out a rule
if not hit for a long time.

The queue bonding mode is default to be disabled, and can be
enabled/disabled with ethtool priv-flags command.

Signed-off-by: Jian Shen <shenjian15@huawei.com>
Signed-off-by: Jijie Shao <shaojijie@huawei.com>
Signed-off-by: Jiantao Xiao <xiaojiantao1@h-partners.com>

K2CI-Arch: Arm64
Link: https://gitee.com/openeuler/kernel/issues/I9A3QT
Signed-off-by: lifangxu <lifangxu@kylinos.cn>
Change-Id: Ie20aa40c3fc270cf921315688b05861c4b53fa12
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/111983
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>