HYGON: KVM: x86: Restore control registers in __set_sregs() to support CSV2 guest live migration
Mainline: KYLIN-only
Severity: Important

task: 216398

When migrate CSV2 guest to the recipient, the KVM which on
recipient's side needs to update the guest context so that the guest
can continues to run. The control register state is necessary for
updating the guest context.

Allows the control registers to be updated in __set_sregs() so that
the CSV2 guest could continue running correctly after migrated to
the recipient.

Signed-off-by: hanliyang <hanliyang@hygon.cn>
Signed-off-by: shitao <shitao@kylinos.cn>
Change-Id: Ieb6594a4e68d553ee6d45d797c358e0be922964c
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/113191
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Reviewed-by: 李唯实 <liweishi@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>