bcachefs: Fix extent merging
Mainline: e320b42dfeb5797a618edbba5186060b3907ba89
From: v6.7-rc1
Severity: Moderate

task: 205710

When merging extents, we have to check that we won't overflow size
fields in any CRC entries - but the check for this was wrong, because in
the loop it was in we weren't keeping a pointer to the (packed, encoded)
CRC field.

Fix this by moving it to its own loop.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>
(backported from commit e320b42dfeb5797a618edbba5186060b3907ba89)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: Ibd9ae888ee345e0d71773e69c55cf8b3f32ec4a5