bcachefs: Debugfs cleanup
Mainline: 9375fbc20079ed0c8a10d8d387b2c173a3dc04d7
From: v6.7-rc1
Severity: Moderate

task: 205710

This improves flush_buf() so that it always returns nonzero when we're
done reading and ready to return to userspace, and so that it returns
the value we want to return to userspace (number of bytes read, if there
wasn't an error).

In the future we'll be better abstracting this mechanism and pulling it
out of bcachefs, and using it to replace seq_file.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>
(backported from commit 9375fbc20079ed0c8a10d8d387b2c173a3dc04d7)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: I83246eab8826daa134560399700c9d7a4f5e76a0