OPENEULER: RDMA/hns: Introduce DCA for RC QP
Mainline: KYLIN-only
Severity: Moderate

task: 219059

The hip09 introduces the DCA(Dynamic context attachment) feature which
supports many RC QPs to share the WQE buffer in a memory pool, this will
reduce the memory consumption when there are too many QPs are inactive.

If a QP enables DCA feature, the WQE's buffer will not be allocated when
creating. But when the users start to post WRs, the hns driver will
allocate a buffer from the memory pool and then fill WQEs which tagged with
this QP's number.

The hns ROCEE will stop accessing the WQE buffer when the user polled all
of the CQEs for a DCA QP, then the driver will recycle this WQE's buffer
to the memory pool.

This patch adds a group of methods to support the user space register
buffers to a memory pool which belongs to the user context. The hns kernel
driver will update the pages state in this pool when the user calling the
post/poll methods and the user driver can get the QP's WQE buffer address
by the key and offset which queried from kernel.

K2CI-Arch: All
Link: https://gitee.com/openeuler/kernel/issues/I9CK0O
Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>
Signed-off-by: lifangxu <lifangxu@kylinos.cn>
Change-Id: I20ff8da66487dedead3da54fd9b271e58bdd658e
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/112651
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>