bcachefs: Rework lru btree
Mainline: 83f33d686553c5105ff36da4dd554c34125094e9
From: v6.7-rc1
Severity: Moderate

task: 205710

This patch changes how the LRU index works:

Instead of using KEY_TYPE_lru where the bucket the lru entry points to
is part of the value, this switches to KEY_TYPE_set and encoding the
bucket we refer to in the low bits of the key.

This means that we no longer have to check for collisions when inserting
LRU entries. We'll be making using of this in the next patch, which adds
a btree write buffer - a pure write buffer for btree updates, where
updates are appended to a simple array and then periodically sorted and
batch inserted.

This is a new on disk format version, and a forced upgrade.

Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>
(backported from commit 83f33d686553c5105ff36da4dd554c34125094e9)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: I5192d2087dfb3567baca872378378ab196c8475c