OPENEULER: net: hns3: add supports pfc storm detection and suppression
Mainline: KYLIN-only
Severity: Moderate

Task: #218871

To solve the problem of abnormal transmission of pauses or PFC frames,
the IMP periodically checks the transmission and reception status of
packets. If a pause frame storm occurs, the pause storm response of
the MAC address is disabled for a period of time.

Signed-off-by: shaojijie <shaojijie@huawei.com>
Signed-off-by: Jiantao Xiao <xiaojiantao1@h-partners.com>

K2CI-Arch: All
Link: https://gitee.com/openeuler/kernel/issues/I97T47
Signed-off-by: Yongzhen Zhang <zhangyongzhen@kylinos.cn>
Change-Id: Ia1765389e651ca8e940bb42b9608952dc8f951dc
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/112140
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>