OPENEULER: net: hns3: add supports fast reporting of faulty nodes
Mainline: KYLIN-only
Severity: Moderate

Task: #218871

Services are switched only when no response is received within 10 seconds.
As a result, services cannot be switched quickly. Therefore, if the chip
is not suspended, the NIC sends a specific message to notify other nodes
of the event. In this way, the service switchover is performed quickly.

Signed-off-by: shaojijie <shaojijie@huawei.com>
Signed-off-by: Jiantao Xiao <xiaojiantao1@h-partners.com>

K2CI-Arch: All
Link: https://gitee.com/openeuler/kernel/issues/I97T47
Signed-off-by: Yongzhen Zhang <zhangyongzhen@kylinos.cn>
Change-Id: I9eabcaaffffa8cb5381c0144556c476ad27b2e7b
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/112141
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>