OPENEULER: net: hns3: add support PF provides customized interfaces to detect port faults.
Mainline: KYLIN-only
Severity: Moderate

Task: #218871

When the link status of the PF network port changes from up to down,
the IMP can quickly report the link down fault cause based on the
link status. If the hardware is faulty from the beginning and the
link is not up, the IMP does not report the link down event
because the status is not changed.

Therefore, an interface is provided to detect port faults.

Signed-off-by: Jie Wang <wangjie125@huawei.com>
Signed-off-by: Jiantao Xiao <xiaojiantao1@h-partners.com>

K2CI-Arch: All
Link: https://gitee.com/openeuler/kernel/issues/I97T47
Signed-off-by: Yongzhen Zhang <zhangyongzhen@kylinos.cn>
Change-Id: I6d0421e6d76fdad7a88179d46f84416ceb919d68
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/112152
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>