accel/habanalabs: prevent immediate hard reset due to 2 adjacent H/W events
Mainline: 5d89ce6f8c2775b926b63c14529c5970429ea935
From: v6.7-rc1
Severity: Moderate

task: 218895

When a H/W event is received while a user is registered to events, no
immediate hard reset will happen, and instead the user will be notified
and will have some time to handle it and eventually release the
device, after which the reset will be done.
If a user, as part of the handling and as part of the cleanup steps
towards releasing the device, unregisters from receiving those events,
and at that time an adjacent H/W event is received, it will be assumed
that the user is not registered to events and thus an immediate hard
reset is required.

To prevent such an unwanted immediate reset, modify the driver to
perform it if the user is not registered to events AND we don't already
have a pending reset for a previous H/W event.

Signed-off-by: Tomer Tayar <ttayar@habana.ai>
Reviewed-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Oded Gabbay <ogabbay@kernel.org>

(backported from commit 5d89ce6f8c2775b926b63c14529c5970429ea935)

K2CI-Arch: All
Signed-off-by: jindongyang <jindongyang@kylinos.cn>
Change-Id: Ib8cb1835bc80f7ad8234b3539fe3df7585396ce5
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/113554
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>