bcachefs: bch2_extent_ptr_desired_durability()
Mainline: 91ecd41b7f02b95279dddcb2193af454efd39497
From: v6.7-rc1
Severity: Moderate

task: 205710

This adds a new helper for getting a pointer's durability irrespective
of the device state, and uses it in the the data update path.

This fixes a bug where we do a data update but request 0 replicas to be
allocated, because the replica being rewritten is on a device marked as
failed.

Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>
(backported from commit 91ecd41b7f02b95279dddcb2193af454efd39497)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: Ic26c2846747bc521fa6ba79f6e007ad00f66c742