wifi: cfg80211: call reg_call_notifier on beacon hints
Mainline: b13b6bbfbb627884f18982600f7b5a5200652531
From: v6.7-rc1
Severity: Moderate

task: 496179

Currently the channel property updates are not propagated to
driver. This causes issues in the discovery of hidden SSIDs and
fails to connect to them.
This change defines a new wiphy flag which when enabled by vendor
driver, the reg_call_notifier callback will be trigger on beacon
hints. This ensures that the channel property changes are visible
to the vendor driver. The vendor changes the channels for active
scans. This fixes the discovery issue of hidden SSID.

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Link: https://lore.kernel.org/r/20230629035254.1.I059fe585f9f9e896c2d51028ef804d197c8c009e@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

K2CI-Arch: All
Change-Id: I9f64cfe259a72646889384c188c1b9b78054664e
Signed-off-by: Yi Cong <yicong@kylinos.cn>
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/91824
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>