io_uring: add IORING_OP_WAITID support
Mainline: f31ecf671ddc498f20219453395794ff2383e06b
From: v6.7-rc1
Severity: Moderate

task: 527473

This adds support for an async version of waitid(2), in a fully async
version. If an event isn't immediately available, wait for a callback
to trigger a retry.

The format of the sqe is as follows:

sqe->len		The 'which', the idtype being queried/waited for.
sqe->fd			The 'pid' (or id) being waited for.
sqe->file_index		The 'options' being set.
sqe->addr2		A pointer to siginfo_t, if any, being filled in.

buf_index, add3, and waitid_flags are reserved/unused for now.
waitid_flags will be used for options for this request type. One
interesting use case may be to add multi-shot support, so that the
request stays armed and posts a notification every time a monitored
process state change occurs.

Note that this does not support rusage, on Arnd's recommendation.

See the waitid(2) man page for details on the arguments.

Signed-off-by: Jens Axboe <axboe@kernel.dk>

(backported from commit f31ecf671ddc498f20219453395794ff2383e06b)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: Icaf998b1a7c7fd9696e1aaa4872b2ec07c7b3d2c
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/114774
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>