accel/habanalabs: set hard reset flag if graceful reset is skipped
Mainline: 4b0b1fbc7757169b6d304545a321c7a88f13f8f0
From: v6.8-rc1
Severity: Moderate

task: 218895

hl_device_cond_reset() might be called with the hard reset flag unset,
because a compute reset upon device release as part of a graceful reset
is valid.
If the conditions for graceful reset are not met, hl_device_reset() will
be called for an immediate reset. In this case a compute reset is not
valid, so it will be replaced with a hard reset together with a debug
message about it.
This message might be confusing, as it implies that a compute reset was
requested when it shouldn't. To prevent this confusion, set the hard
reset flag in hl_device_cond_reset() if going to an immediate reset.

Signed-off-by: Tomer Tayar <ttayar@habana.ai>
Reviewed-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Oded Gabbay <ogabbay@kernel.org>

(backported from commit 4b0b1fbc7757169b6d304545a321c7a88f13f8f0)

K2CI-Arch: All
Signed-off-by: jindongyang <jindongyang@kylinos.cn>
Change-Id: I9822ca8839107cc348853ea82e763cd10a84ca6e
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/114188
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>