bcachefs: Optimize bch2_dirent_name_bytes
Mainline: 29c336afebb9013706334c1b451f31fe5ff37f34
From: v6.7-rc1
Severity: Moderate

task: 205710

Avoids doing a full strnlen for getting the length of the name of a
dirent entry.

Given the fact that the name of dirents is stored at the end of the
bkey's value, and we know the length of that in u64s, we can find the
last u64 and figure out how many NUL bytes are at the end of the string.

On little endian systems this ends up being the leading zeros of the
last u64, whereas on big endian systems this ends up being the trailing
zeros of the last u64.
We can take that value in bits and divide it by 8 to get the number of
NUL bytes at the end.

There is no endian-fixup or other compatibility here as this is string
data interpreted as a u64.

Signed-off-by: Joshua Ashton <joshua@froggi.es>
Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>
(backported from commit 29c336afebb9013706334c1b451f31fe5ff37f34)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: I080ea4c3c821da655376f1d8ad639b54d35e3b24