ovl: Support escaped overlay.* xattrs
Mainline: dad02fad84cbce30f317b69a4f2391f90045f79d
From: v6.7-rc1
Severity: Important

task: 508351

There are cases where you want to use an overlayfs mount as a lowerdir
for another overlayfs mount. For example, if the system rootfs is on
overlayfs due to composefs, or to make it volatile (via tmps), then
you cannot currently store a lowerdir on the rootfs. This means you
can't e.g. store on the rootfs a prepared container image for use
using overlayfs.

To work around this, we introduce an escapment mechanism for overlayfs
xattrs. Whenever the lower/upper dir has a xattr named
"overlay.overlay.XYZ", we list it as "overlay.XYZ" in listxattrs, and
when the user calls getxattr or setxattr on "overlay.XYZ", we apply to
"overlay.overlay.XYZ" in the backing directories.

This allows storing any kind of overlay xattrs in a overlayfs mount
that can be used as a lowerdir in another mount. It is possible to
stack this mechanism multiple times, such that
"overlay.overlay.overlay.XYZ" will survive two levels of overlay mounts,
however this is not all that useful in practice because of stack depth
limitations of overlayfs mounts.

Note: These escaped xattrs are copied to upper during copy-up.

Signed-off-by: Alexander Larsson <alexl@redhat.com>
Reviewed-by: Amir Goldstein <amir73il@gmail.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>

(backported from commit dad02fad84cbce30f317b69a4f2391f90045f79d)

Change-Id: If1ead18d1f60321359d9cf54855d0a80641adb41
Signed-off-by: kuangkai <kuangkai@kylinos.cn>
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/95552
Reviewed-by: 杜晋秀 <dujinxiu@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 刘云 <liuyun01@kylinos.cn>