iommu: Remove useless group refcounting
Mainline: e946f8e3e62bf05da21a14658f8cb05e2a616260
From: v6.7-rc1
Severity: Moderate

task: 218851

Several functions obtain the group reference and then release it before
returning. This gives the impression that the refcount is protecting
something for the duration of the function.

In truth all of these functions are called in places that know a device
driver is probed to the device and our locking rules already require
that dev->iommu_group cannot change while a driver is attached to the
struct device.

If this was not the case then this code is already at risk of triggering
UAF as it is racy if the dev->iommu_group is concurrently going to
NULL/free. refcount debugging will throw a WARN if kobject_get() is
called on a 0 refcount object to highlight the bug.

Remove the confusing refcounting and leave behind a comment about the
restriction.

Reviewed-by: Lu Baolu <baolu.lu@linux.intel.com>
Reviewed-by: Kevin Tian <kevin.tian@intel.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Link: https://lore.kernel.org/r/1-v1-c869a95191f2+5e8-iommu_single_grp_jgg@nvidia.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>
(backported from commit e946f8e3e62bf05da21a14658f8cb05e2a616260)

K2CI-Arch: All
Signed-off-by: lifangxu <lifangxu@kylinos.cn>
Change-Id: I2e5b0cbe94073e614caa612a3b77c0ee266b51b5
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/111212
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>