ovl: Add an alternative type of whiteout
Mainline: bc8df7a3dc035903426a17ea3027f55817de13a8
From: v6.7-rc1
Severity: Important

task: 508351

An xattr whiteout (called "xwhiteout" in the code) is a reguar file of
zero size with the "overlay.whiteout" xattr set. A file like this in a
directory with the "overlay.whiteouts" xattrs set will be treated the
same way as a regular whiteout.

The "overlay.whiteouts" directory xattr is used in order to
efficiently handle overlay checks in readdir(), as we only need to
checks xattrs in affected directories.

The advantage of this kind of whiteout is that they can be escaped
using the standard overlay xattr escaping mechanism. So, a file with a
"overlay.overlay.whiteout" xattr would be unescaped to
"overlay.whiteout", which could then be consumed by another overlayfs
as a whiteout.

Overlayfs itself doesn't create whiteouts like this, but a userspace
mechanism could use this alternative mechanism to convert images that
may contain whiteouts to be used with overlayfs.

To work as a whiteout for both regular overlayfs mounts as well as
userxattr mounts both the "user.overlay.whiteout*" and the
"trusted.overlay.whiteout*" xattrs will need to be created.

Signed-off-by: Alexander Larsson <alexl@redhat.com>
Reviewed-by: Amir Goldstein <amir73il@gmail.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>

(backported from commit bc8df7a3dc035903426a17ea3027f55817de13a8)

Change-Id: I29257e0913d66821419f635994f5857adeceb2c4
Signed-off-by: kuangkai <kuangkai@kylinos.cn>
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/95553
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 杜晋秀 <dujinxiu@kylinos.cn>
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>